Updated README's roadmap to reflect changes/Main.go has new functionallity/Database file Profiles.txt added
main.go now has the functionality to pull data from the database file Profiles.txt with grouped lines. It also has the limited ability to add some profile's information back into the file. Questioning the purpose of the profile struct, when the database can be altered and saved directly. More thought needs to be given to requirement of having the struct.