Logging Node Revamped as a connection Server
Logging node now allows remote connections and loggs all trafic sent to it.
It still retains the manual search and pull from log.txt's to help with its versitility.
UI has been reworked to allow the setting of ports in the applications.
This is done without a config file because I want these to be launched multiple times remotely.